requirements.txt
torch==2.0.1
transformers==4.28.1
hydra-core==1.3.2
mlflow==2.10.2
scikit-learn==1.2.2
datasets==2.17.0
evaluate==0.4.0